feat:fixed getUseSelector
