+++ src/index.js
@@ -4,8 +4,8 @@ import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import "./i18next";
-import { Provider } from "react-redux";
 import { store } from "./redux/store";
+import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import "./firebase";
 import "slick-carousel/slick/slick.css";
